+++ src/Entity/Support/Rdv.php
@@ -9,6 +9,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Mapping\Annotation as Gedmo;
 use Gedmo\SoftDeleteable\Traits\SoftDeleteableEntity;
+use Symfony\Component\Serializer\Annotation\Groups;
 use Symfony\Component\Validator\Constraints as Assert;
 
 /**
@@ -33,50 +34,59 @@ class Rdv
      * @ORM\Id()
      * @ORM\GeneratedValue()
      * @ORM\Column(type="integer")
+     * @Groups("show_rdv")
      */
     private $id;
 
     /**
      * @ORM\Column(type="string", length=255, nullable=false)
      * @Assert\NotBlank()
+     * @Groups("show_rdv")
      */
     private $title;
 
     /**
      * @ORM\Column(type="datetime")
      * @Assert\NotNull()
+     * @Groups("show_rdv")
      */
     private $start;
 
     /**
      * @ORM\Column(type="datetime")
      * @Assert\NotNull()
+     * @Groups("show_rdv")
      */
     private $end;
 
     /**
      * @ORM\Column(type="smallint", nullable=true)
+     * @Groups("show_rdv")
      */
     private $status;
 
     /**
      * @ORM\Column(type="string", length=255, nullable=true)
+     * @Groups("show_rdv")
      */
     private $location;
 
     /**
      * @ORM\Column(type="text", nullable=true)
+     * @Groups("show_rdv")
      */
     private $content;
 
     /**
      * @Gedmo\Blameable(on="create")
      * @ORM\ManyToOne(targetEntity="App\Entity\Organization\User", inversedBy="rdvs")
+     * @Groups("show_rdv")
      */
     protected $createdBy; // NE PAS SUPPRIMER
 
     /**
      * @ORM\ManyToOne(targetEntity="App\Entity\Support\SupportGroup", inversedBy="rdvs", fetch="EXTRA_LAZY")
+     * @Groups("show_rdv")
      */
     private $supportGroup;
 
@@ -84,6 +94,18 @@ class Rdv
      * @ORM\ManyToOne(targetEntity=User::class, inversedBy="rdvs2")
      */
     private $user;
+
+    /**
+     * @ORM\Column(type="string", nullable=true)
+     * @Groups("show_rdv")
+     */
+    private $googleEventId;
+
+    /**
+     * @ORM\Column(type="string", length=255, nullable=true)
+     * @Groups("show_rdv")
+     */
+    private $outlookEventId;
 
     public function __construct()
     {
@@ -192,6 +214,30 @@ class Rdv
     public function setUser(?User $user): self
     {
         $this->user = $user;
+
+        return $this;
+    }
+
+    public function getGoogleEventId(): ?string
+    {
+        return $this->googleEventId;
+    }
+
+    public function setGoogleEventId(?string $googleEventId): self
+    {
+        $this->googleEventId = $googleEventId;
+
+        return $this;
+    }
+
+    public function getOutlookEventId(): ?string
+    {
+        return $this->outlookEventId;
+    }
+
+    public function setOutlookEventId(?string $outlookEventId): self
+    {
+        $this->outlookEventId = $outlookEventId;
 
         return $this;
     }